% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ebola_gamma_pstr.R
\docType{data}
\name{bets_results_bootstrap}
\alias{bets_results_bootstrap}
\title{Alternative bootstrapped parameter estimates of COVID-19 incubation period from Zhao et al.}
\format{
A data frame with 1,000 rows and 6 variables:
\itemize{
  \item \code{shape} shape parameter of gamma distribution
  \item \code{rate} rate parameter of gamma distribution
  \item \code{idx} index of sample
  \item \code{median} the median correponding to this bootstrapped sample
  \item \code{p95} the 95th percentile correponding to this bootstrapped sample
  }
}
\usage{
data(bets_results_bootstrap)
}
\description{
Alternative bootstrapped parameter estimates of COVID-19 incubation period from Zhao et al.
}
\references{
\itemize{
  \item Zhao, Q., Ju N., Bacallado, S., & Shah, R. D. BETS: The dangers of selection bias in early analyses of the coronavirus disease (COVID-19) pandemic. arXiv (2020).
}
}
