% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ebola_gamma_pstr.R
\docType{data}
\name{boot_lnorm_params_covid}
\alias{boot_lnorm_params_covid}
\title{Bootstrapped parameter estimates of COVID-19 incubation period}
\format{
A data frame with 1,000 rows and 5 variables:
\itemize{
  \item \code{meanlog} meanlog parameter of log-normal distribution
  \item \code{sdlog} sdlog parameter of log-normal distribution
  \item \code{idx} index of sample
  \item \code{median} the median correponding to this bootstrapped sample
  \item \code{p95} the 95th percentile correponding to this bootstrapped sample
  }
}
\usage{
data(boot_lnorm_params_covid)
}
\description{
A dataset containing samples from an estimated bootstrap distribution of the
incubation period of COVID-19 was estimated from data
collected and stored at https://github.com/HopkinsIDD/ncov_incubation.
}
\references{
\itemize{
  \item Lauer, S. A., Grantz, K. H., Bi, Q., Jones, F. K., Zheng, Q., Meredith, H. R., Azman, A. S., Reich, N. G., & Lessler, J. The incubation period of coronavirus disease 2019 (COVID-19) from publicly reported confirmed cases: estimation and application. Annals of Internal Medicine 172 (9), 577-582 (2020).
}
}
