% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ebola_gamma_pstr.R
\docType{data}
\name{boot_lnorm_params_ncov}
\alias{boot_lnorm_params_ncov}
\title{Bootstrapped parameter estimates of 2019-nCoV incubation period}
\format{A data frame with 1,000 rows and 2 variables:
\itemize{
  \item \code{meanlog} meanlog parameter of log-normal distribution
  \item \code{sdlog} sdlog parameter of log-normal distribution
  \item \code{idx} index of sample
  \item \code{median} the median correponding to this bootstrapped sample
  \item \code{p95} the 95th percentile correponding to this bootstrapped sample
  }}
\usage{
data(boot_lnorm_params_ncov)
}
\description{
A dataset containing samples from an estimated bootstrap distribution of the
incubation period of 2019-nCoV was estimated from data
collected and stored at https://github.com/HopkinsIDD/ncov_incubation.
}
\references{
\itemize{
  \item Lauer, S. A., Grantz, K. H., Bi, Q., Jones, F. K., Zheng, Q., Meredith, H., Azman, A. S., Reich, N. G., & Lessler, J. The incubation period of 2019-nCoV from publicly reported confirmed cases: estimation and application. medRxiv (2020).
}
}
