% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calc_monitoring_costs.R
\name{calc_monitoring_costs}
\alias{calc_monitoring_costs}
\title{Cost calculations for active monitoring scenarios}
\usage{
calc_monitoring_costs(durs, probs_of_disease, per_day_hazard, N, cost_mat,
  gamma_params, return_scalar = FALSE)
}
\arguments{
\item{durs}{a vector of AM durations to consider, in multiples of median incubation period values}

\item{probs_of_disease}{a vector of probabilities representing the chances that different subgroups have of developing the disease of interest}

\item{per_day_hazard}{hazard to pass to the calculation of theta}

\item{N}{number of indiviuals for which to calculate the costs}

\item{cost_mat}{4x2 matrix of costs where each named row corresponds to a cost, and the columns are the lower and upper bounds}

\item{gamma_params}{the median, shape and scale parameters for the assumed-gamma incubation period distribution}

\item{return_scalar}{logical, whether to return the scalar max cost value using the first value of durs and probs_of_disease}
}
\value{
a length(probs_of_disease)xlength(durs) matrix of costs, one row per subgroup, one column per duration
}
\description{
Cost calculations for active monitoring scenarios
}

