% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ebola_gamma_pstr.R
\docType{data}
\name{kde_smallpox}
\alias{kde_smallpox}
\title{Kernel Density Estimates for posterior distributions}
\format{
Output from fit_kde() function in inst/analysis-code/inc-per-mcmc.R. Data used in plotting credible regions of posterior estimates.
}
\usage{
data(kde_ebola)
}
\description{
Kernel Density Estimates for posterior distributions
}
