% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ebola_gamma_pstr.R
\docType{data}
\name{mers_gamma_pstr}
\alias{mers_gamma_pstr}
\title{Posterior distribution of MERS-CoV incubation period}
\format{A data frame with 5 variables:
\itemize{
  \item \code{shape} shape parameter of gamma distribution
  \item \code{scale} scale parameter of gamma distribution
  \item \code{idx} index of sample
  \item \code{median} the median correponding to this posterior sample
  \item \code{p95} the 95th percentile correponding to this posterior sample
  }}
\usage{
data(mers_gamma_pstr)
}
\description{
A dataset containing samples from an estimated posterior distribution of the 
incubation period of MERS-CoV. Distribution was estimated from de-identified
data from Virlogeux et al. (2016).
}
\references{
Virlogeux, V., Park, M., Wu, J. T. & Cowling, B. J. Association between Severity of MERS-CoV Infection and Incubation Period. Emerging Infectious Diseases 22, (2016).
}

