% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_risk.R
\name{plot_modified_credible_regions}
\alias{plot_modified_credible_regions}
\title{make posterior likelihood plot for gamma distribution}
\usage{
plot_modified_credible_regions(
  mcmc_dfs,
  kdes,
  label_txt,
  colors,
  show.legend = FALSE,
  base.size = 12
)
}
\arguments{
\item{mcmc_dfs}{the rbound outputs from inc_per_mcmc}

\item{kdes}{list of results from call to fit_kde}

\item{label_txt}{labels for points}

\item{colors}{colors for each region}

\item{show.legend}{logical, passed to geom_point}

\item{base.size}{size for font, to be passed to theme_bw()}
}
\description{
make posterior likelihood plot for gamma distribution
}
