% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_risk.R
\name{plot_risk}
\alias{plot_risk}
\title{Function for plotting risk of case escaping active monitoring}
\usage{
plot_risk(
  pstr_data,
  u = 7,
  phi = c(1e-04, 0.001, 0.01),
  durations = 5:25,
  return_plot = FALSE
)
}
\arguments{
\item{pstr_data}{a data object with posterior distributions of gamma parameters (must have shape and scale column)}

\item{u}{numeric value or vector of assumed duration(s) of time between exposure and monitoring}

\item{phi}{probability a case becomes symptomatic}

\item{durations}{durations of active monitoring to plot}

\item{return_data}{logical, whether or not to return the data used for plotting}
}
\value{
if return_plot is specified, it returns the grob
}
\description{
Function for plotting risk of case escaping active monitoring
}
\examples{
plot_risk(ebola_gamma_pstr, u=c(5, 10))
}
