% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_risk.R
\name{plot_risk_gdist}
\alias{plot_risk_gdist}
\title{Function for plotting risk and uncertainty of case escaping active monitoring}
\usage{
plot_risk_gdist(
  dist = "gamma",
  arg_list = list(shape = 5.807, scale = 0.948),
  nsamp = 1000,
  u,
  udist = "unif",
  uargs = list(min = 1, max = 14),
  phi = c(1e-04, 0.001, 0.01),
  durations = 5:25,
  ci_width = 0.9,
  yrange = NULL,
  include_xlab = TRUE,
  include_legend = TRUE,
  output_plot = TRUE,
  return_data = FALSE,
  return_plot = FALSE
)
}
\arguments{
\item{dist}{character string specifying the parametric distribution (as in \code{\link[stats]{distributions}}), defaults to \code{"gamma"}}

\item{arg_list}{list of named numeric vectors to be plugged into distribution in \code{dist}}

\item{nsamp}{number of samples from posterior distribution upon which to base calculations}

\item{u}{numeric value or vector of assumed duration(s) of time between exposure and monitoring}

\item{udist}{character string specifying the parameteric distribution of assumed durations of time between exposure and monitoring (if missing \code{u}), defaults to \code{"unif"} (Uniform distribution)}

\item{uargs}{list of named numeric vectors to be plugged into distribution in \code{udist}, defaults to \code{list(min=1, max=14)}}

\item{phi}{probability a case becomes symptomatic}

\item{durations}{durations of active monitoring to plot}

\item{ci_width}{numeric value between 0 and 1 indicating the nominal value for the eventually computed and displayed CI}

\item{yrange}{if not NULL, a vector of the ylim to plot}

\item{include_xlab}{logical, whether to include an x-axis label}

\item{include_legend}{logical, whether to include a legend}

\item{output_plot}{logical, whether or not to automatically output the plot}

\item{return_data}{logical, whether or not to return the data used for plotting}

\item{return_plot}{logical, whether or not to return the plotted grob}
}
\value{
if return_plot is specified, it returns the grob
}
\description{
Function for plotting risk and uncertainty of case escaping active monitoring
}
\examples{
## load Ebola data and run
data(pstr_gamma_params_ebola)
plot_risk_gdist(dist="gamma",
                arg_list=list(shape=pstr_gamma_params_ebola$shape,
                              scale=pstr_gamma_params_ebola$scale))

}
