% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_of_missing_case.R
\name{prob_of_missing_case}
\alias{prob_of_missing_case}
\title{Calculate probabilility of a case not being caught by monitoring}
\usage{
prob_of_missing_case(dat, dist = "gamma")
}
\arguments{
\item{dat}{a data.frame with column names explained in details}

\item{dist}{character string specifying the parametric distribution of dat}
}
\value{
the inputted data.frame with an additional column called 'p'
}
\description{
Calculate probabilility of a case not being caught by monitoring
}
\details{
Column names of the dat argument
\itemize{
 \item{"phi"}{the probability of an exposed individual becoming symptomatic}
 \item{"shape"}{shape parameter of the gamma distribution for incubation periods}
 \item{"scale"}{scale parameter of the gamma distribution for incubation periods}
 \item{"d"}{the duration of the active monitoring period}
 \item{"u"}{the assumed number of days between exposure and the beginning of monitoring}
}
}
