% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ebola_gamma_pstr.R
\docType{data}
\name{pstr_gamma_params_mers}
\alias{pstr_gamma_params_mers}
\title{Posterior distribution of MERS-CoV incubation period}
\format{
A data frame with 1,000,000 rows and 6 variables:
\itemize{
  \item \code{shape} shape parameter of gamma distribution
  \item \code{scale} scale parameter of gamma distribution
  \item \code{idx} index of sample
  \item \code{median} the median correponding to this posterior sample
  \item \code{p95} the 95th percentile correponding to this posterior sample
  \item \code{chain} which chain the sample belongs to
  }
}
\usage{
data(pstr_gamma_params_mers)
}
\description{
A dataset containing samples from an estimated posterior distribution of the
incubation period of MERS-CoV. Distribution was estimated from de-identified
data from Virlogeux et al. (2016).
}
\references{
Virlogeux, V., Park, M., Wu, J. T. & Cowling, B. J. Association between Severity of MERS-CoV Infection and Incubation Period. Emerging Infectious Diseases 22, (2016).
}
