% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ebola_gamma_pstr.R
\docType{data}
\name{pstr_gamma_params_smallpox}
\alias{pstr_gamma_params_smallpox}
\title{Posterior distribution of smallpox incubation period}
\format{
A data frame with 1,000,000 rows and 6 variables:
\itemize{
  \item \code{shape} shape parameter of gamma distribution
  \item \code{scale} scale parameter of gamma distribution
  \item \code{idx} index of sample
  \item \code{median} the median correponding to this posterior sample
  \item \code{p95} the 95th percentile correponding to this posterior sample
  \item \code{chain} which chain the sample belongs to
  }
}
\usage{
data(pstr_gamma_params_smallpox)
}
\description{
A dataset containing samples from an estimated posterior distribution of the
incubation period of smallpox. Distribution was estimated from de-identified
data from several sources (see References).
}
\references{
\itemize{
  \item Litvinjenko, S., Arsic, B. & Borjanovic, S. Epidemiologic Aspects of Smallpox in Yugoslavia in 1972. Bulletin of the World Health Organization (1972).
  \item Mack, T. M. Smallpox in Europe, 1950-1971. Journal of Infectious Diseases 125, 161–169 (1972).
  \item Nishiura, H. Determination of the appropriate quarantine period following smallpox exposure: an objective approach using the incubation period distribution. Int J Hyg Environ Health 212, 97–104 (2009).
}
}
