% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_risk.R
\name{plot_risk_uncertainty}
\alias{plot_risk_uncertainty}
\title{Function for plotting risk and uncertainty of case escaping active monitoring}
\usage{
plot_risk_uncertainty(pstr_data, max_u = 14, phi = c(1e-04, 0.001, 0.01),
  durations = 5:25, yrange = NULL, include_xlab = TRUE,
  include_legend = TRUE, return_plot = FALSE)
}
\arguments{
\item{pstr_data}{a data object with posterior distributions of gamma parameters (must have shape and scale column)}

\item{phi}{probability a case becomes symptomatic}

\item{durations}{durations of active monitoring to plot}

\item{yrange}{if not NULL, a vector of the ylim to plot}

\item{include_xlab}{logical, whether to include an x-axis label}

\item{include_legend}{logical, whether to include a legend}

\item{u}{numeric value or vector of assumed duration(s) of time between exposure and monitoring}

\item{return_data}{logical, whether or not to return the data used for plotting}
}
\value{
if return_plot is specified, it returns the grob
}
\description{
Function for plotting risk and uncertainty of case escaping active monitoring
}
\examples{
plot_risk_uncertainty(ebola_gamma_pstr)
}

